Move `Buffer` to `buffer` module
