+++ vulkano/src/buffer/sys.rs
@@ -13,35 +13,19 @@
 //! other buffer types of this library, and all custom buffer types
 //! that you create must wrap around the types in this module.
 
-use super::{
-    cpu_access::{ReadLockError, WriteLockError},
-    BufferContents, BufferCreateFlags, BufferUsage,
-};
+use super::{Buffer, BufferCreateFlags, BufferError, BufferMemory, BufferUsage};
 use crate::{
     device::{Device, DeviceOwned},
     memory::{
-        allocator::{
-            AllocationCreationError, AllocationType, DeviceAlignment, DeviceLayout, MemoryAlloc,
-        },
-        DedicatedTo, ExternalMemoryHandleType, ExternalMemoryHandleTypes, MemoryAllocateFlags,
-        MemoryPropertyFlags, MemoryRequirements,
+        allocator::{AllocationType, DeviceAlignment, DeviceLayout, MemoryAlloc},
+        DedicatedTo, ExternalMemoryHandleTypes, MemoryAllocateFlags, MemoryPropertyFlags,
+        MemoryRequirements,
     },
-    range_map::RangeMap,
-    sync::{future::AccessError, CurrentAccess, Sharing},
-    DeviceSize, RequirementNotMet, RequiresOneOf, Version, VulkanError, VulkanObject,
+    sync::Sharing,
+    DeviceSize, RequiresOneOf, Version, VulkanError, VulkanObject,
 };
-use parking_lot::{Mutex, MutexGuard};
 use smallvec::SmallVec;
-use std::{
-    error::Error,
-    fmt::{Display, Error as FmtError, Formatter},
-    hash::{Hash, Hasher},
-    mem::{size_of_val, MaybeUninit},
-    num::NonZeroU64,
-    ops::{Deref, DerefMut, Range},
-    ptr,
-    sync::Arc,
-};
+use std::{mem::MaybeUninit, num::NonZeroU64, ptr, sync::Arc};
 
 /// A raw buffer, with no memory backing it.
 ///
@@ -607,8 +591,8 @@ impl RawBuffer {
 
     /// Returns the usage the buffer was created with.
     #[inline]
-    pub fn usage(&self) -> &BufferUsage {
-        &self.usage
+    pub fn usage(&self) -> BufferUsage {
+        self.usage
     }
 
     /// Returns the sharing the buffer was created with.
@@ -652,7 +636,7 @@ unsafe impl DeviceOwned for RawBuffer {
 
 crate::impl_id_counter!(RawBuffer);
 
-/// Parameters to create a new `Buffer`.
+/// Parameters to create a new [`Buffer`].
 #[derive(Clone, Debug)]
 pub struct BufferCreateInfo {
     /// Flags to enable.
@@ -698,739 +682,6 @@ impl Default for BufferCreateInfo {
             usage: BufferUsage::empty(),
             external_memory_handle_types: ExternalMemoryHandleTypes::empty(),
             _ne: crate::NonExhaustive(()),
-        }
-    }
-}
-
-/// A storage for raw bytes.
-///
-/// Unlike [`RawBuffer`], a `Buffer` has memory backing it, and can be used normally.
-#[derive(Debug)]
-pub struct Buffer {
-    inner: RawBuffer,
-    memory: BufferMemory,
-    state: Mutex<BufferState>,
-}
-
-/// The type of backing memory that a buffer can have.
-#[derive(Debug)]
-pub enum BufferMemory {
-    /// The buffer is backed by normal memory, bound with [`bind_memory`].
-    ///
-    /// [`bind_memory`]: RawBuffer::bind_memory
-    Normal(MemoryAlloc),
-
-    /// The buffer is backed by sparse memory, bound with [`bind_sparse`].
-    ///
-    /// [`bind_sparse`]: crate::device::QueueGuard::bind_sparse
-    Sparse,
-}
-
-impl Buffer {
-    fn from_raw(inner: RawBuffer, memory: BufferMemory) -> Self {
-        let state = Mutex::new(BufferState::new(inner.size));
-
-        Buffer {
-            inner,
-            memory,
-            state,
-        }
-    }
-
-    /// Returns the type of memory that is backing this buffer.
-    #[inline]
-    pub fn memory(&self) -> &BufferMemory {
-        &self.memory
-    }
-
-    /// Returns the memory requirements for this buffer.
-    #[inline]
-    pub fn memory_requirements(&self) -> &MemoryRequirements {
-        &self.inner.memory_requirements
-    }
-
-    /// Returns the flags the buffer was created with.
-    #[inline]
-    pub fn flags(&self) -> BufferCreateFlags {
-        self.inner.flags
-    }
-
-    /// Returns the size of the buffer in bytes.
-    #[inline]
-    pub fn size(&self) -> DeviceSize {
-        self.inner.size
-    }
-
-    /// Returns the usage the buffer was created with.
-    #[inline]
-    pub fn usage(&self) -> &BufferUsage {
-        &self.inner.usage
-    }
-
-    /// Returns the sharing the buffer was created with.
-    #[inline]
-    pub fn sharing(&self) -> &Sharing<SmallVec<[u32; 4]>> {
-        &self.inner.sharing
-    }
-
-    /// Returns the external memory handle types that are supported with this buffer.
-    #[inline]
-    pub fn external_memory_handle_types(&self) -> ExternalMemoryHandleTypes {
-        self.inner.external_memory_handle_types
-    }
-
-    /// Locks the buffer in order to read its content from the host.
-    ///
-    /// If the buffer is currently used in exclusive mode by the device, this function will return
-    /// an error. Similarly if you called `write()` on the buffer and haven't dropped the lock,
-    /// this function will return an error as well.
-    ///
-    /// After this function successfully locks the buffer, any attempt to submit a command buffer
-    /// that uses it in exclusive mode will fail. You can still submit this buffer for non-exclusive
-    /// accesses (ie. reads).
-    pub fn read(&self, range: Range<DeviceSize>) -> Result<BufferReadGuard<'_, [u8]>, BufferError> {
-        assert!(!range.is_empty() && range.end <= self.inner.size);
-
-        let allocation = match &self.memory {
-            BufferMemory::Normal(a) => a,
-            BufferMemory::Sparse => todo!("`Buffer::read` doesn't support sparse binding yet"),
-        };
-
-        if allocation.mapped_ptr().is_none() {
-            return Err(BufferError::MemoryNotHostVisible);
-        }
-
-        let mut state = self.state();
-
-        unsafe {
-            state.check_cpu_read(range.clone())?;
-            state.cpu_read_lock(range.clone());
-        }
-
-        let data = unsafe {
-            // If there are other read locks being held at this point, they also called
-            // `invalidate_range` when locking. The GPU can't write data while the CPU holds a read
-            // lock, so there will be no new data and this call will do nothing.
-            // TODO: probably still more efficient to call it only if we're the first to acquire a
-            // read lock, but the number of CPU locks isn't currently tracked anywhere.
-            allocation.invalidate_range(0..self.size()).unwrap();
-            allocation.mapped_slice().unwrap()
-        };
-
-        Ok(BufferReadGuard {
-            buffer: self,
-            range,
-            data,
-        })
-    }
-
-    /// Locks the buffer in order to write its content from the host.
-    ///
-    /// If the buffer is currently in use by the device, this function will return an error.
-    /// Similarly if you called `read()` on the buffer and haven't dropped the lock, this function
-    /// will return an error as well.
-    ///
-    /// After this function successfully locks the buffer, any attempt to submit a command buffer
-    /// that uses it and any attempt to call `read()` will return an error.
-    pub fn write(&self, range: Range<DeviceSize>) -> Result<BufferWriteGuard<'_>, BufferError> {
-        assert!(!range.is_empty() && range.end <= self.inner.size);
-
-        let allocation = match &self.memory {
-            BufferMemory::Normal(a) => a,
-            BufferMemory::Sparse => todo!("`Buffer::write` doesn't support sparse binding yet"),
-        };
-
-        if allocation.mapped_ptr().is_none() {
-            return Err(BufferError::MemoryNotHostVisible);
-        }
-
-        let mut state = self.state();
-
-        unsafe {
-            state.check_cpu_write(range.clone())?;
-            state.cpu_write_lock(range.clone());
-        }
-
-        let data = unsafe {
-            allocation.invalidate_range(0..self.size()).unwrap();
-            allocation.write(0..self.size()).unwrap()
-        };
-
-        Ok(BufferWriteGuard {
-            buffer: self,
-            range,
-            data,
-        })
-    }
-
-    pub(crate) fn state(&self) -> MutexGuard<'_, BufferState> {
-        self.state.lock()
-    }
-}
-
-unsafe impl VulkanObject for Buffer {
-    type Handle = ash::vk::Buffer;
-
-    #[inline]
-    fn handle(&self) -> Self::Handle {
-        self.inner.handle
-    }
-}
-
-unsafe impl DeviceOwned for Buffer {
-    #[inline]
-    fn device(&self) -> &Arc<Device> {
-        &self.inner.device
-    }
-}
-
-impl PartialEq for Buffer {
-    #[inline]
-    fn eq(&self, other: &Self) -> bool {
-        self.inner == other.inner
-    }
-}
-
-impl Eq for Buffer {}
-
-impl Hash for Buffer {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.inner.hash(state);
-    }
-}
-
-/// The current state of a buffer.
-#[derive(Debug)]
-pub(crate) struct BufferState {
-    ranges: RangeMap<DeviceSize, BufferRangeState>,
-}
-
-impl BufferState {
-    fn new(size: DeviceSize) -> Self {
-        BufferState {
-            ranges: [(
-                0..size,
-                BufferRangeState {
-                    current_access: CurrentAccess::Shared {
-                        cpu_reads: 0,
-                        gpu_reads: 0,
-                    },
-                },
-            )]
-            .into_iter()
-            .collect(),
-        }
-    }
-
-    pub(crate) fn check_cpu_read(&self, range: Range<DeviceSize>) -> Result<(), ReadLockError> {
-        for (_range, state) in self.ranges.range(&range) {
-            match &state.current_access {
-                CurrentAccess::CpuExclusive { .. } => return Err(ReadLockError::CpuWriteLocked),
-                CurrentAccess::GpuExclusive { .. } => return Err(ReadLockError::GpuWriteLocked),
-                CurrentAccess::Shared { .. } => (),
-            }
-        }
-
-        Ok(())
-    }
-
-    pub(crate) unsafe fn cpu_read_lock(&mut self, range: Range<DeviceSize>) {
-        self.ranges.split_at(&range.start);
-        self.ranges.split_at(&range.end);
-
-        for (_range, state) in self.ranges.range_mut(&range) {
-            match &mut state.current_access {
-                CurrentAccess::Shared { cpu_reads, .. } => {
-                    *cpu_reads += 1;
-                }
-                _ => unreachable!("Buffer is being written by the CPU or GPU"),
-            }
-        }
-    }
-
-    pub(crate) unsafe fn cpu_read_unlock(&mut self, range: Range<DeviceSize>) {
-        self.ranges.split_at(&range.start);
-        self.ranges.split_at(&range.end);
-
-        for (_range, state) in self.ranges.range_mut(&range) {
-            match &mut state.current_access {
-                CurrentAccess::Shared { cpu_reads, .. } => *cpu_reads -= 1,
-                _ => unreachable!("Buffer was not locked for CPU read"),
-            }
-        }
-    }
-
-    pub(crate) fn check_cpu_write(&self, range: Range<DeviceSize>) -> Result<(), WriteLockError> {
-        for (_range, state) in self.ranges.range(&range) {
-            match &state.current_access {
-                CurrentAccess::CpuExclusive => return Err(WriteLockError::CpuLocked),
-                CurrentAccess::GpuExclusive { .. } => return Err(WriteLockError::GpuLocked),
-                CurrentAccess::Shared {
-                    cpu_reads: 0,
-                    gpu_reads: 0,
-                } => (),
-                CurrentAccess::Shared { cpu_reads, .. } if *cpu_reads > 0 => {
-                    return Err(WriteLockError::CpuLocked)
-                }
-                CurrentAccess::Shared { .. } => return Err(WriteLockError::GpuLocked),
-            }
-        }
-
-        Ok(())
-    }
-
-    pub(crate) unsafe fn cpu_write_lock(&mut self, range: Range<DeviceSize>) {
-        self.ranges.split_at(&range.start);
-        self.ranges.split_at(&range.end);
-
-        for (_range, state) in self.ranges.range_mut(&range) {
-            state.current_access = CurrentAccess::CpuExclusive;
-        }
-    }
-
-    pub(crate) unsafe fn cpu_write_unlock(&mut self, range: Range<DeviceSize>) {
-        self.ranges.split_at(&range.start);
-        self.ranges.split_at(&range.end);
-
-        for (_range, state) in self.ranges.range_mut(&range) {
-            match &mut state.current_access {
-                CurrentAccess::CpuExclusive => {
-                    state.current_access = CurrentAccess::Shared {
-                        cpu_reads: 0,
-                        gpu_reads: 0,
-                    }
-                }
-                _ => unreachable!("Buffer was not locked for CPU write"),
-            }
-        }
-    }
-
-    pub(crate) fn check_gpu_read(&self, range: Range<DeviceSize>) -> Result<(), AccessError> {
-        for (_range, state) in self.ranges.range(&range) {
-            match &state.current_access {
-                CurrentAccess::Shared { .. } => (),
-                _ => return Err(AccessError::AlreadyInUse),
-            }
-        }
-
-        Ok(())
-    }
-
-    pub(crate) unsafe fn gpu_read_lock(&mut self, range: Range<DeviceSize>) {
-        self.ranges.split_at(&range.start);
-        self.ranges.split_at(&range.end);
-
-        for (_range, state) in self.ranges.range_mut(&range) {
-            match &mut state.current_access {
-                CurrentAccess::GpuExclusive { gpu_reads, .. }
-                | CurrentAccess::Shared { gpu_reads, .. } => *gpu_reads += 1,
-                _ => unreachable!("Buffer is being written by the CPU"),
-            }
-        }
-    }
-
-    pub(crate) unsafe fn gpu_read_unlock(&mut self, range: Range<DeviceSize>) {
-        self.ranges.split_at(&range.start);
-        self.ranges.split_at(&range.end);
-
-        for (_range, state) in self.ranges.range_mut(&range) {
-            match &mut state.current_access {
-                CurrentAccess::GpuExclusive { gpu_reads, .. } => *gpu_reads -= 1,
-                CurrentAccess::Shared { gpu_reads, .. } => *gpu_reads -= 1,
-                _ => unreachable!("Buffer was not locked for GPU read"),
-            }
-        }
-    }
-
-    pub(crate) fn check_gpu_write(&self, range: Range<DeviceSize>) -> Result<(), AccessError> {
-        for (_range, state) in self.ranges.range(&range) {
-            match &state.current_access {
-                CurrentAccess::Shared {
-                    cpu_reads: 0,
-                    gpu_reads: 0,
-                } => (),
-                _ => return Err(AccessError::AlreadyInUse),
-            }
-        }
-
-        Ok(())
-    }
-
-    pub(crate) unsafe fn gpu_write_lock(&mut self, range: Range<DeviceSize>) {
-        self.ranges.split_at(&range.start);
-        self.ranges.split_at(&range.end);
-
-        for (_range, state) in self.ranges.range_mut(&range) {
-            match &mut state.current_access {
-                CurrentAccess::GpuExclusive { gpu_writes, .. } => *gpu_writes += 1,
-                &mut CurrentAccess::Shared {
-                    cpu_reads: 0,
-                    gpu_reads,
-                } => {
-                    state.current_access = CurrentAccess::GpuExclusive {
-                        gpu_reads,
-                        gpu_writes: 1,
-                    }
-                }
-                _ => unreachable!("Buffer is being accessed by the CPU"),
-            }
-        }
-    }
-
-    pub(crate) unsafe fn gpu_write_unlock(&mut self, range: Range<DeviceSize>) {
-        self.ranges.split_at(&range.start);
-        self.ranges.split_at(&range.end);
-
-        for (_range, state) in self.ranges.range_mut(&range) {
-            match &mut state.current_access {
-                &mut CurrentAccess::GpuExclusive {
-                    gpu_reads,
-                    gpu_writes: 1,
-                } => {
-                    state.current_access = CurrentAccess::Shared {
-                        cpu_reads: 0,
-                        gpu_reads,
-                    }
-                }
-                CurrentAccess::GpuExclusive { gpu_writes, .. } => *gpu_writes -= 1,
-                _ => unreachable!("Buffer was not locked for GPU write"),
-            }
-        }
-    }
-}
-
-/// The current state of a specific range of bytes in a buffer.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-struct BufferRangeState {
-    current_access: CurrentAccess,
-}
-
-/// RAII structure used to release the CPU read access of a buffer when dropped.
-///
-/// This structure is created by the [`read`] method on [`Buffer`].
-///
-/// [`read`]: Buffer::read
-#[derive(Debug)]
-pub struct BufferReadGuard<'a, T>
-where
-    T: BufferContents + ?Sized,
-{
-    buffer: &'a Buffer,
-    range: Range<DeviceSize>,
-    data: &'a T,
-}
-
-impl<'a, T> Drop for BufferReadGuard<'a, T>
-where
-    T: BufferContents + ?Sized + 'a,
-{
-    fn drop(&mut self) {
-        unsafe {
-            let mut state = self.buffer.state();
-            state.cpu_read_unlock(self.range.clone());
-        }
-    }
-}
-
-impl<'a, T> Deref for BufferReadGuard<'a, T>
-where
-    T: BufferContents + ?Sized + 'a,
-{
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        self.data
-    }
-}
-
-/// RAII structure used to release the CPU write access of a buffer when dropped.
-///
-/// This structure is created by the [`write`] method on [`Buffer`].
-///
-/// [`write`]: Buffer::write
-#[derive(Debug)]
-pub struct BufferWriteGuard<'a> {
-    buffer: &'a Buffer,
-    range: Range<DeviceSize>,
-    data: &'a mut [u8],
-}
-
-impl<'a> Drop for BufferWriteGuard<'a> {
-    fn drop(&mut self) {
-        let allocation = match &self.buffer.memory {
-            BufferMemory::Normal(a) => a,
-            BufferMemory::Sparse => unreachable!(),
-        };
-
-        unsafe {
-            allocation.flush_range(0..self.buffer.size()).unwrap();
-
-            let mut state = self.buffer.state();
-            state.cpu_write_unlock(self.range.clone());
-        }
-    }
-}
-
-impl<'a> Deref for BufferWriteGuard<'a> {
-    type Target = [u8];
-
-    fn deref(&self) -> &Self::Target {
-        self.data
-    }
-}
-
-impl<'a> DerefMut for BufferWriteGuard<'a> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.data
-    }
-}
-
-/// Error that can happen in buffer functions.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum BufferError {
-    VulkanError(VulkanError),
-
-    /// Allocating memory failed.
-    AllocError(AllocationCreationError),
-
-    RequirementNotMet {
-        required_for: &'static str,
-        requires_one_of: RequiresOneOf,
-    },
-
-    /// The memory was created dedicated to a resource, but not to this buffer.
-    DedicatedAllocationMismatch,
-
-    /// A dedicated allocation is required for this buffer, but one was not provided.
-    DedicatedAllocationRequired,
-
-    /// The host is already using this buffer in a way that is incompatible with the
-    /// requested access.
-    InUseByHost,
-
-    /// The device is already using this buffer in a way that is incompatible with the
-    /// requested access.
-    InUseByDevice,
-
-    /// The specified size exceeded the value of the `max_buffer_size` limit.
-    MaxBufferSizeExceeded {
-        size: DeviceSize,
-        max: DeviceSize,
-    },
-
-    /// The offset of the allocation does not have the required alignment.
-    MemoryAllocationNotAligned {
-        allocation_offset: DeviceSize,
-        required_alignment: DeviceAlignment,
-    },
-
-    /// The size of the allocation is smaller than what is required.
-    MemoryAllocationTooSmall {
-        allocation_size: DeviceSize,
-        required_size: DeviceSize,
-    },
-
-    /// The buffer was created with the `shader_device_address` usage, but the memory does not
-    /// support this usage.
-    MemoryBufferDeviceAddressNotSupported,
-
-    /// The memory was created with export handle types, but none of these handle types were
-    /// enabled on the buffer.
-    MemoryExternalHandleTypesDisjoint {
-        buffer_handle_types: ExternalMemoryHandleTypes,
-        memory_export_handle_types: ExternalMemoryHandleTypes,
-    },
-
-    /// The memory was created with an import, but the import's handle type was not enabled on
-    /// the buffer.
-    MemoryImportedHandleTypeNotEnabled {
-        buffer_handle_types: ExternalMemoryHandleTypes,
-        memory_imported_handle_type: ExternalMemoryHandleType,
-    },
-
-    /// The memory backing this buffer is not visible to the host.
-    MemoryNotHostVisible,
-
-    /// The protection of buffer and memory are not equal.
-    MemoryProtectedMismatch {
-        buffer_protected: bool,
-        memory_protected: bool,
-    },
-
-    /// The provided memory type is not one of the allowed memory types that can be bound to this
-    /// buffer.
-    MemoryTypeNotAllowed {
-        provided_memory_type_index: u32,
-        allowed_memory_type_bits: u32,
-    },
-
-    /// The sharing mode was set to `Concurrent`, but one of the specified queue family indices was
-    /// out of range.
-    SharingQueueFamilyIndexOutOfRange {
-        queue_family_index: u32,
-        queue_family_count: u32,
-    },
-}
-
-impl Error for BufferError {
-    fn source(&self) -> Option<&(dyn Error + 'static)> {
-        match self {
-            BufferError::VulkanError(err) => Some(err),
-            BufferError::AllocError(err) => Some(err),
-            _ => None,
-        }
-    }
-}
-
-impl Display for BufferError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), FmtError> {
-        match self {
-            Self::VulkanError(_) => write!(f, "a runtime error occurred"),
-            Self::AllocError(_) => write!(f, "allocating memory failed"),
-            Self::RequirementNotMet {
-                required_for,
-                requires_one_of,
-            } => write!(
-                f,
-                "a requirement was not met for: {}; requires one of: {}",
-                required_for, requires_one_of,
-            ),
-            Self::DedicatedAllocationMismatch => write!(
-                f,
-                "the memory was created dedicated to a resource, but not to this buffer",
-            ),
-            Self::DedicatedAllocationRequired => write!(
-                f,
-                "a dedicated allocation is required for this buffer, but one was not provided"
-            ),
-            Self::InUseByHost => write!(
-                f,
-                "the host is already using this buffer in a way that is incompatible with the \
-                requested access",
-            ),
-            Self::InUseByDevice => write!(
-                f,
-                "the device is already using this buffer in a way that is incompatible with the \
-                requested access"
-            ),
-            Self::MaxBufferSizeExceeded { .. } => write!(
-                f,
-                "the specified size exceeded the value of the `max_buffer_size` limit",
-            ),
-            Self::MemoryAllocationNotAligned {
-                allocation_offset,
-                required_alignment,
-            } => write!(
-                f,
-                "the offset of the allocation ({}) does not have the required alignment ({:?})",
-                allocation_offset, required_alignment,
-            ),
-            Self::MemoryAllocationTooSmall {
-                allocation_size,
-                required_size,
-            } => write!(
-                f,
-                "the size of the allocation ({}) is smaller than what is required ({})",
-                allocation_size, required_size,
-            ),
-            Self::MemoryBufferDeviceAddressNotSupported => write!(
-                f,
-                "the buffer was created with the `shader_device_address` usage, but the memory \
-                does not support this usage",
-            ),
-            Self::MemoryExternalHandleTypesDisjoint { .. } => write!(
-                f,
-                "the memory was created with export handle types, but none of these handle types \
-                were enabled on the buffer",
-            ),
-            Self::MemoryImportedHandleTypeNotEnabled { .. } => write!(
-                f,
-                "the memory was created with an import, but the import's handle type was not \
-                enabled on the buffer",
-            ),
-            Self::MemoryNotHostVisible => write!(
-                f,
-                "the memory backing this buffer is not visible to the host",
-            ),
-            Self::MemoryProtectedMismatch {
-                buffer_protected,
-                memory_protected,
-            } => write!(
-                f,
-                "the protection of buffer ({}) and memory ({}) are not equal",
-                buffer_protected, memory_protected,
-            ),
-            Self::MemoryTypeNotAllowed {
-                provided_memory_type_index,
-                allowed_memory_type_bits,
-            } => write!(
-                f,
-                "the provided memory type ({}) is not one of the allowed memory types (",
-                provided_memory_type_index,
-            )
-            .and_then(|_| {
-                let mut first = true;
-
-                for i in (0..size_of_val(allowed_memory_type_bits))
-                    .filter(|i| allowed_memory_type_bits & (1 << i) != 0)
-                {
-                    if first {
-                        write!(f, "{}", i)?;
-                        first = false;
-                    } else {
-                        write!(f, ", {}", i)?;
-                    }
-                }
-
-                Ok(())
-            })
-            .and_then(|_| write!(f, ") that can be bound to this buffer")),
-            Self::SharingQueueFamilyIndexOutOfRange { .. } => write!(
-                f,
-                "the sharing mode was set to `Concurrent`, but one of the specified queue family \
-                indices was out of range",
-            ),
-        }
-    }
-}
-
-impl From<VulkanError> for BufferError {
-    fn from(err: VulkanError) -> BufferError {
-        match err {
-            VulkanError::OutOfHostMemory => BufferError::AllocError(
-                AllocationCreationError::VulkanError(VulkanError::OutOfHostMemory),
-            ),
-            VulkanError::OutOfDeviceMemory => BufferError::AllocError(
-                AllocationCreationError::VulkanError(VulkanError::OutOfDeviceMemory),
-            ),
-            _ => panic!("unexpected error: {:?}", err),
-        }
-    }
-}
-
-impl From<RequirementNotMet> for BufferError {
-    fn from(err: RequirementNotMet) -> Self {
-        Self::RequirementNotMet {
-            required_for: err.required_for,
-            requires_one_of: err.requires_one_of,
-        }
-    }
-}
-
-impl From<ReadLockError> for BufferError {
-    fn from(err: ReadLockError) -> Self {
-        match err {
-            ReadLockError::CpuWriteLocked => Self::InUseByHost,
-            ReadLockError::GpuWriteLocked => Self::InUseByDevice,
-        }
-    }
-}
-
-impl From<WriteLockError> for BufferError {
-    fn from(err: WriteLockError) -> Self {
-        match err {
-            WriteLockError::CpuLocked => Self::InUseByHost,
-            WriteLockError::GpuLocked => Self::InUseByDevice,
         }
     }
 }
